Cat refactored into header files
