+++ Wrappers.cpp
@@ -1,0 +1,13 @@
+#include <iostream>
+#include "Wrappers.h"
+
+CatWrapper::CatWrapper(Cat* ptrToCat) : catPtr(ptrToCat) 
+{ 
+    std::cout << "+++ CatWrapper +++" << std::endl;
+}
+
+CatWrapper::~CatWrapper()
+{
+    delete catPtr;
+    std::cout << "/// CatWrapper ///" << std::endl;
+}
