+++ Wrappers.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "Cat.h"
+
+struct CatWrapper
+{
+    Cat* catPtr = nullptr;
+    CatWrapper(Cat* ptrToCat); 
+    ~CatWrapper();
+};
