+++ main.cpp
@@ -61,126 +61,11 @@ void Axe::aConstMemberFunction() const { }
          The file that a nested type is defined in MUST be included wherever that nested type is written.
  */
 
-
-
-
-
 #include <iostream>
 #include <iomanip>
 #include "LeakedObjectDetector.h"
-
-/*
- copied UDT 1:
- */
-struct Cat
-{
-    Cat();
-    ~Cat();
-    int numLimbs;
-    int numTails;
-    float ageYears;
-    std::string colour;
-    float weightKg;
-    int cutenessLevel;
-    int maxCroquettes;
-
-    void meow(int loudnessDB);
-
-    bool catchAnimal(std::string animalSpecies); 
-
-    float throwUpFurBall(); 
-
-    int eatCroquettes(int croquettesNum, float croquetteWeight);
-
-    JUCE_LEAK_DETECTOR(Cat)
-};
-
-Cat::Cat() :
-numLimbs(3),
-numTails(1),
-ageYears(3.6f),
-colour("grey"),
-weightKg(4.1f),
-cutenessLevel(50),
-maxCroquettes(7)
-{
-    std::cout << "Cat is born" << std::endl;
-}
-
-Cat::~Cat() 
-{
-    if (cutenessLevel < 0)
-    {
-        std::cout << "Cat is gone to cat hell" << std::endl;
-    }
-    else
-    {
-        std::cout << "Cat is gone to cat heaven" << std::endl;
-    }
-}
-
-
-void Cat::meow(int loudnessDB)
-{
-    std::cout << "Meow!!! " << loudnessDB << " dB" << std::endl;
-    if (loudnessDB < 30)
-    {
-        cutenessLevel += 10;
-    }
-    else
-    {
-        cutenessLevel -= loudnessDB;        
-    }
-    std::cout << "My current cuteness level is " << cutenessLevel << std::endl;
-}
-
-bool Cat::catchAnimal(std::string animalSpecies)
-{
-    if (animalSpecies == "Fly")
-    {
-        cutenessLevel += 10;
-        return true;
-    }
-    cutenessLevel -= 10;
-    return false;
-}
-
-float Cat::throwUpFurBall()
-{
-    cutenessLevel -= 50;
-    return weightKg * 11;
-}
-
-int Cat::eatCroquettes(int croquettesNum, float croquetteWeightG)
-{
-    int i = 0;
-    while (i < croquettesNum)
-    {
-        ++i;
-        weightKg += croquetteWeightG * 0.001f;
-        std::cout << "Eating a croquette... yum!\n";
-        if (i >= maxCroquettes) 
-        {
-            std::cout << "I had enough.\n";
-            return croquettesNum - i;
-        }
-    }
-    return 0;
-}
-
-struct CatWrapper
-{
-    Cat* catPtr = nullptr;
-    CatWrapper(Cat* ptrToCat) : catPtr(ptrToCat) 
-    { 
-        std::cout << "+++ CatWrapper +++" << std::endl;
-    }
-    ~CatWrapper()
-    {
-        delete catPtr;
-        std::cout << "/// CatWrapper ///" << std::endl;
-    }
-};
+#include "Cat.h"
+#include "Wrappers.h"
 
 /*
  copied UDT 2:
